Revert "RavenDB-15919 adjust subscription behaviour"
This reverts commit 08131c9be55053b806c2ac96da00a03d84e5683f.
