+++ src/Raven.Server/Documents/Subscriptions/SubscriptionConnectionsState.cs
@@ -199,6 +199,34 @@ namespace Raven.Server.Documents.Subscriptions
             }
         }
 
+        public class ResendItem
+        {
+            public string Id;
+            public long Batch;
+            public string ChangeVector;
+            public SubscriptionType Type;
+        }
+
+        public IEnumerable<ResendItem> GetResendItems(ClusterOperationContext context)
+        {
+            var subscriptionState = context.Transaction.InnerTransaction.OpenTable(ClusterStateMachine.SubscriptionStateSchema, ClusterStateMachine.SubscriptionState);
+            using (GetDatabaseAndSubscriptionPrefix(context, DocumentDatabase.Name, SubscriptionId, out var prefix))
+            using (Slice.External(context.Allocator, prefix, out var prefixSlice))
+            {
+                var resendItem = new ResendItem();
+             
+                foreach (var item in subscriptionState.SeekByPrimaryKeyPrefix(prefixSlice, Slices.Empty, 0))
+                {
+                    resendItem.Type = (SubscriptionType)item.Key[prefixSlice.Size];
+                    resendItem.Id = item.Value.Reader.ReadStringWithPrefix((int)ClusterStateMachine.SubscriptionStateTable.Key, prefix.Length + 2);
+                    resendItem.ChangeVector = item.Value.Reader.ReadString((int)ClusterStateMachine.SubscriptionStateTable.ChangeVector);
+                    resendItem.Batch = Bits.SwapBytes(item.Value.Reader.ReadLong((int)ClusterStateMachine.SubscriptionStateTable.BatchId));
+
+                    yield return resendItem;
+                }
+            }
+        }
+
         public Task AcknowledgeBatch(SubscriptionConnection connection, long batchId, List<DocumentRecord> addDocumentsToResend)
         {
             if (ClusterCommandsVersionManager.CurrentClusterMinimalVersion >= 53_000)
@@ -435,7 +463,7 @@ namespace Raven.Server.Documents.Subscriptions
         {
             if (_connections.TryRemove(connection))
             {
-                CancelSingleConnection(connection, ex ?? connection.ConnectionException ?? new SubscriptionClosedException($"Connection {connection.ClientUri} has closed", canReconnect: true));
+                CancelSingleConnection(connection, ex ?? new SubscriptionClosedException($"Connection {connection.ClientUri} has closed", canReconnect: true));
                 NotifyHasMoreDocs(); //upon connection fail, all recorded docs should be resent
             }
         }
