+++ test/SlowTests/Client/Subscriptions/RavenDB_9512.cs
@@ -39,8 +39,7 @@ namespace SlowTests.Client.Subscriptions
                 
                 var st = worker.Run(x => { });
                 
-                Assert.True(await Assert.ThrowsAsync<SubscriptionClosedException>(()=>st).WaitAsync(_reasonableWaitTime));
-            }
+                Assert.True(await Assert.ThrowsAsync<SubscriptionClosedException>(()=>st).WaitAsync(_reasonableWaitTime));}
         }
         
     }
